move sqlmodels to database module and merge Season model with SeasonTorrents model
