+++ MediaManager/src/dowloadClients/genericDownloadClient.py
@@ -1,4 +1,4 @@
-from dowloadClients import Torrents
+from database.tv import Season
 
 
 class GenericDownloadClient(object):
@@ -10,7 +10,7 @@ class GenericDownloadClient(object):
         self.name = name
 
     # TODO: change Torrents type to SeasonTorrents|MovieTorrents
-    def download(self, torrent: Torrents) -> Torrents:
+    def download(self, torrent: Season) -> Season:
         """
         downloads a torrent
 
@@ -18,7 +18,7 @@ class GenericDownloadClient(object):
         """
         raise NotImplementedError()
 
-    def get_torrent_status(self, torrent: Torrents) -> Torrents:
+    def get_torrent_status(self, torrent: Season) -> Season:
         """
         updates a torrents 'status' field
